# Conflicts:
#	OUIKit/OUIApplet/src/main/java/io/openim/android/ouiapplet/vm/AppletVM.java
#	OUIKit/OUIGroup/src/main/java/io/openim/android/ouigroup/ui/GroupMaterialActivity.java

+++ OUIKit/OUICore/src/main/java/io/openim/android/ouicore/entity/NotificationContent.java
@@ -9,7 +9,7 @@ public class NotificationContent implements Serializable {
 
     public String ownerUserID;
 
-    public int createTime;
+    public long createTime;
 
     public int memberCount;
